refactor: optimized box-component for map-building
