+++ src/ions/levels/level-0.js
@@ -27,6 +27,7 @@ createLevel({
 			position: [0, -2.5, 0],
 			args: [30, 3, 3],
 			rotation: [0, 0, 0],
+			axis: [null, null, null],
 			color: colors.primaryDark,
 			metalness: "0.3",
 			roughness: "0.5",
@@ -36,6 +37,7 @@ createLevel({
 			position: [0, 16.875, 0],
 			args: [30, 3, 3],
 			rotation: [0, 0, 0],
+			axis: [null, null, null],
 			color: colors.primaryDark,
 			metalness: "0.3",
 			roughness: "0.5",
@@ -45,6 +47,7 @@ createLevel({
 			position: [-13.5, 6, 0],
 			args: [3, 19, 3],
 			rotation: [0, 0, 0],
+			axis: [null, null, null],
 			color: colors.primaryDark,
 			metalness: "0.3",
 			roughness: "0.5",
@@ -54,6 +57,7 @@ createLevel({
 			position: [13.5, 6, 0],
 			args: [3, 19, 3],
 			rotation: [0, 0, 0],
+			axis: [null, null, null],
 			color: colors.primaryDark,
 			metalness: "0.3",
 			roughness: "0.5",
@@ -63,6 +67,7 @@ createLevel({
 			position: [0, 7.5, -3],
 			args: [30, 20, 3],
 			rotation: [0, 0, 0],
+			axis: [null, null, null],
 			color: colors.primaryMain,
 			metalness: "0.3",
 			roughness: "0.5",
@@ -72,6 +77,7 @@ createLevel({
 			position: [0, 11, 0],
 			args: [5, 10, 3],
 			rotation: [0, 0, 0],
+			axis: [null, null, null],
 			color: colors.primaryDark,
 			metalness: "0.3",
 			roughness: "0.5",
