+++ src/ions/levels/level-3.js
@@ -27,6 +27,7 @@ createLevel({
 			position: [0, -2.5, 0],
 			args: [30, 3, 3],
 			rotation: [0, 0, 0],
+			axis: [null, null, null],
 			color: colors.primaryDark,
 			metalness: "0.3",
 			roughness: "0.5",
@@ -36,6 +37,7 @@ createLevel({
 			position: [0, 16.875, 0],
 			args: [30, 3, 3],
 			rotation: [0, 0, 0],
+			axis: [null, null, null],
 			color: colors.primaryDark,
 			metalness: "0.3",
 			roughness: "0.5",
@@ -45,6 +47,7 @@ createLevel({
 			position: [-13.5, 6, 0],
 			args: [3, 19, 3],
 			rotation: [0, 0, 0],
+			axis: [null, null, null],
 			color: colors.primaryDark,
 			metalness: "0.3",
 			roughness: "0.5",
@@ -54,6 +57,7 @@ createLevel({
 			position: [13.5, 6, 0],
 			args: [3, 19, 3],
 			rotation: [0, 0, 0],
+			axis: [null, null, null],
 			color: colors.primaryDark,
 			metalness: "0.3",
 			roughness: "0.5",
@@ -63,6 +67,7 @@ createLevel({
 			position: [0, 7.5, -3],
 			args: [30, 20, 3],
 			rotation: [0, 0, 0],
+			axis: [null, null, null],
 			color: colors.primaryMain,
 			metalness: "0.3",
 			roughness: "0.5",
@@ -72,6 +77,7 @@ createLevel({
 			position: [1, 2, -1],
 			args: [10, 6, 3],
 			rotation: [2, 0, 0],
+			axis: [null, null, null],
 			color: "#22a6b3",
 			metalness: "0.3",
 			roughness: "0.5",
@@ -81,6 +87,7 @@ createLevel({
 			position: [-6, 12, 1],
 			args: [8, 4, 3],
 			rotation: [0, 2, 0],
+			axis: [null, null, null],
 			color: "#22a6b3",
 			metalness: "0.3",
 			roughness: "0.5",
@@ -90,6 +97,7 @@ createLevel({
 			position: [6, 9, 0],
 			args: [4, 2, 3],
 			rotation: [0, 0, 2],
+			axis: [null, null, null],
 			color: "#22a6b3",
 			metalness: "0.3",
 			roughness: "0.5",
